requirements.txt
fastapi
uvicorn
pymongo
paho-mqtt
requests
jinja2
